export FusionAuthLoginButton from tested component and consolidate
