+++ src/components/index.ts
@@ -1,4 +1,4 @@
-import FusionAuthLoginButton from "./FusionAuthLoginButton.vue";
+import FusionAuthLoginButton from "./FusionAuthLoginButton/FusionAuthLoginButton.vue";
 import FusionAuthLogoutButton from "./FusionAuthLogoutButton.vue";
 import FusionAuthRegisterButton from "./FusionAuthRegisterButton.vue";
 import RequireAuth from "./RequireAuth.vue";
